ajustes

+++ atak-teste-api/src/main/java/testeatakapi/testeatakapi/seguranca/dominio/UsuarioRepositorio.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepositorio extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepositorio extends JpaRepository<Usuario, Integer> {
 
     Optional<Usuario> findByUsuario(String usuario);
